\name{RFRE.PO-package}
\alias{RFRE.PO-package}
\alias{RFRE.PO}
\docType{package}
\title{
\packageTitle{RFRE.PO}
}
\description{
\packageDescription{RFRE.PO}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{RFRE.PO}

Maintainer: \packageMaintainer{RFRE.PO}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = RFRE.PO)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{RFRE.PO}}
%\section{Documentation Index}{\packageIndices{RFRE.PO}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
